pyproject.toml
[project]
name = "secure-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography-fernet>=0.1.0",
    "streamlit>=1.45.0",
]
